% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndi_to_csv.R
\name{ndi_to_csv}
\alias{ndi_to_csv}
\title{Reads and converts an Optotrak file from given filepath and re-saves as a CSV file}
\usage{
ndi_to_csv(ndi_path, output_path, sep = ",")
}
\description{
Reads and converts an Optotrak file from given filepath and re-saves as a CSV file
}
