% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ndi.R
\name{read_ndi}
\alias{read_ndi}
\title{Reads and converts an Optotrak file from given filepath into a opto object}
\usage{
read_ndi(filepath)
}
\description{
Reads and converts an Optotrak file from given filepath into a opto object
}
